basic function has been met.

+++ Backend/api/nft.go
@@ -5,10 +5,10 @@ import (
 )
 
 type NFTRepository interface {
-	CreateNFT(tokenID, contractAddress string, ownerID, creatorID int64, metadataURI string) (int64, error)
-	GetNFTByID(id int64) (*models.NFT, error)
+	CreateNFT(tokenID, contractAddress string, ownerID, creatorID uint, metadataURI string) (uint, error)
+	GetNFTByID(id uint) (*models.NFT, error)
 	GetNFTByTokenID(tokenID string) (*models.NFT, error)
 	UpdateNFT(nft *models.NFT) error
-	DeleteNFT(id int64) error
-	GetNFTsByOwnerID(ownerID int64) ([]*models.NFT, error)
+	DeleteNFT(id uint) error
+	GetNFTsByOwnerID(ownerID uint) ([]*models.NFT, error)
 }
