+++ Backend/api/transaction.go
@@ -5,9 +5,9 @@ import (
 )
 
 type TransactionRepository interface {
-	CreateTransaction(orderID int64, txHash string, amount, gasFee float64) (int64, error)
-	GetTransactionByID(id int64) (*models.Transaction, error)
+	CreateTransaction(orderID uint, txHash string, amount, gasFee float64) (uint, error)
+	GetTransactionByID(id uint) (*models.Transaction, error)
 	GetTransactionByTxHash(txHash string) (*models.Transaction, error)
-	UpdateTransactionStatus(id int64, status string) error
-	GetTransactionsByOrderID(orderID int64) ([]*models.Transaction, error)
+	UpdateTransactionStatus(id uint, status string) error
+	GetTransactionsByOrderID(orderID uint) ([]*models.Transaction, error)
 }
