+++ Backend/api/user_balance.go
@@ -1,12 +1,12 @@
 package api
 
 import (
-	_"nftPlantform/models"
+	_ "nftPlantform/models"
 )
 
 type UserBalanceRepository interface {
-	GetBalance(userID int64) (float64, error)
-	UpdateBalance(userID int64, newBalance float64) error
-	AddToBalance(userID int64, amount float64) error
-	SubtractFromBalance(userID int64, amount float64) error
+	GetBalance(userID uint) (float64, error)
+	UpdateBalance(userID uint, newBalance float64) error
+	AddToBalance(userID uint, amount float64) error
+	SubtractFromBalance(userID uint, amount float64) error
 }
